Moves many parameters their own header file
This commit creates a new header file `parameters.hh`, which contains
an equally named namespace with two sub-namespaces `numerical` and
`physical`. These contain a lot of different parameters, such as the
grid sizes and iteration goals.
The advantage of having them in a separate file like this, is, that they
only need to be changed at one central location instead of all over the
code, reducing the risk of accidentally missing a place when changing
some of them.